nat: extract destination-NAT tests into tests_destination.rs (#4409)
Move the 40 destination NAT (DNAT) tests -- dnat_basic_lookup_tcp and
the source-scoped, prefix/longest-match, multiple-destination, zone
precedence, port-aware and snat-merge cases -- verbatim into
tests_destination.rs, together with the test-local dnat_row() helper
(used only by the prefix-destination tests in this group).

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 118 tests.
